marathon_planner_agent/requirements.txt
google-adk>=1.28.1
google-genai>=1.70.0
